# student-growth-system-design-v3/student-growth-system-design-v3/api/openapi.yaml
openapi: 3.1.0
info: {title: Student Growth Archive API, version: 3.4.0, license: {name: MIT, identifier: MIT}}
servers: [{url: /api/v1}]
security: [{bearerAuth: []}]
paths:
  /students:
    get: {operationId: listStudents, summary: 查询学生列表, responses: {'200': {$ref: '#/components/responses/StudentList'}, '401': {$ref: '#/components/responses/Unauthorized'}, '500': {$ref: '#/components/responses/InternalServerError'}}}
    post:
      operationId: createStudent
      summary: 新增学生
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/StudentCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/StudentOne'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
  /students/{studentId}:
    parameters: [{$ref: '#/components/parameters/StudentId'}]
    get: {operationId: getStudent, summary: 查询学生详情, responses: {'200': {$ref: '#/components/responses/StudentOne'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '500': {$ref: '#/components/responses/InternalServerError'}}}
    put:
      operationId: updateStudent
      summary: 修改学生
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/StudentUpdate'}}}}
      responses: {'200': {$ref: '#/components/responses/StudentOne'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
  /exams:
    post:
      operationId: createExam
      summary: 新增考试及成绩
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ExamCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/ExamOne'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
  /exams/{examId}:
    parameters: [{$ref: '#/components/parameters/ExamId'}]
    get: {operationId: getExam, summary: 查询考试, responses: {'200': {$ref: '#/components/responses/ExamOne'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '500': {$ref: '#/components/responses/InternalServerError'}}}
    put:
      operationId: updateExam
      summary: 修改考试
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ExamUpdate'}}}}
      responses: {'200': {$ref: '#/components/responses/ExamOne'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
    delete: {operationId: deleteExam, summary: 删除考试, responses: {'204': {description: 删除成功}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '500': {$ref: '#/components/responses/InternalServerError'}}}
  /wrong-questions:
    get:
      operationId: listWrongQuestions
      summary: 分页查询错题
      parameters: [{$ref: '#/components/parameters/StudentIdQuery'}, {$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PageSize'}]
      responses: {'200': {$ref: '#/components/responses/WrongPage'}, '401': {$ref: '#/components/responses/Unauthorized'}, '500': {$ref: '#/components/responses/InternalServerError'}}
    post:
      operationId: createWrongQuestion
      summary: 新增错题
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/WrongCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/WrongOne'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
  /wrong-questions/{wrongQuestionId}:
    parameters: [{$ref: '#/components/parameters/WrongId'}]
    get: {operationId: getWrongQuestion, summary: 查询错题, responses: {'200': {$ref: '#/components/responses/WrongOne'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '500': {$ref: '#/components/responses/InternalServerError'}}}
    put:
      operationId: updateWrongQuestion
      summary: 修改错题
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/WrongUpdate'}}}}
      responses: {'200': {$ref: '#/components/responses/WrongOne'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
    delete: {operationId: deleteWrongQuestion, summary: 删除错题, responses: {'204': {description: 删除成功}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '500': {$ref: '#/components/responses/InternalServerError'}}}
  /wrong-questions/{wrongQuestionId}/reviews:
    parameters: [{$ref: '#/components/parameters/WrongId'}, {$ref: '#/components/parameters/IdempotencyKey'}]
    post:
      operationId: submitWrongQuestionReview
      summary: 提交复习结果
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ReviewCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/ReviewOne'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
  /reviews/due:
    get:
      operationId: listDueReviews
      summary: 查询到期复习
      parameters: [{name: studentId, in: query, required: true, schema: {$ref: '#/components/schemas/Id'}}, {name: before, in: query, schema: {type: string, format: date-time}}, {$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PageSize'}]
      responses: {'200': {$ref: '#/components/responses/WrongPage'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
  /resources:
    get:
      operationId: listResources
      summary: 分页查询资源
      parameters: [{$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PageSize'}]
      responses: {'200': {$ref: '#/components/responses/ResourcePage'}, '401': {$ref: '#/components/responses/Unauthorized'}, '500': {$ref: '#/components/responses/InternalServerError'}}
    post:
      operationId: createResource
      summary: 新增资源
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ResourceCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/ResourceOne'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
  /resources/{resourceId}:
    parameters: [{$ref: '#/components/parameters/ResourceId'}]
    get: {operationId: getResource, summary: 查询资源, responses: {'200': {$ref: '#/components/responses/ResourceOne'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '500': {$ref: '#/components/responses/InternalServerError'}}}
    put:
      operationId: updateResource
      summary: 修改资源
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ResourceUpdate'}}}}
      responses: {'200': {$ref: '#/components/responses/ResourceOne'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
  /resources/{resourceId}/history:
    post:
      operationId: createResourceHistory
      summary: 新增资源学习历史
      parameters: [{$ref: '#/components/parameters/ResourceId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/ResourceHistoryCreateRequest'}}}}
      responses: {'201': {$ref: '#/components/responses/ResourceHistoryOne'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
    get:
      operationId: listResourceHistory
      summary: 查询资源学习历史
      parameters: [{$ref: '#/components/parameters/ResourceId'}, {name: studentId, in: query, required: true, schema: {$ref: '#/components/schemas/Id'}}, {$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PageSize'}]
      responses: {'200': {$ref: '#/components/responses/ResourceHistoryPage'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
  /resource-history:
    get:
      operationId: listStudentResourceHistory
      summary: 查询学生资源学习历史
      parameters: [{name: studentId, in: query, required: true, schema: {$ref: '#/components/schemas/Id'}}, {name: subjectId, in: query, schema: {$ref: '#/components/schemas/Id'}}, {name: resourceType, in: query, schema: {type: string}}, {name: sourceType, in: query, schema: {type: string}}, {name: completed, in: query, schema: {type: boolean}}, {name: startDate, in: query, schema: {type: string, format: date}}, {name: endDate, in: query, schema: {type: string, format: date}}, {$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PageSize'}]
      responses: {'200': {$ref: '#/components/responses/ResourceHistoryPage'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
  /student-resources:
    get:
      operationId: listStudentResources
      summary: 分页查询学生资源队列
      parameters: [{name: studentId, in: query, required: true, schema: {$ref: '#/components/schemas/Id'}}, {name: status, in: query, schema: {$ref: '#/components/schemas/StudentResourceStatus'}}, {name: subjectId, in: query, schema: {$ref: '#/components/schemas/Id'}}, {$ref: '#/components/parameters/Page'}, {$ref: '#/components/parameters/PageSize'}]
      responses: {'200': {$ref: '#/components/responses/StudentResourcePage'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
    post:
      operationId: createStudentResource
      summary: 将资源加入学生资源队列
      description: student 必须存在，resource 必须存在且未逻辑删除；同一 student + resource 只能有一条 assignment，重复创建返回 409 DUPLICATE_DATA。
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/StudentResourceCreateRequest'}}}}
      responses: {'201': {$ref: '#/components/responses/StudentResourceOne'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}
...
  /stages:
    get: {operationId: listStages, summary: 查询学段, parameters: [{name: stageId, in: query, schema: {$ref: '#/components/schemas/Id'}}, {name: enabledOnly, in: query, schema: {type: boolean, default: true}}], responses: {'200': {$ref: '#/components/responses/StageList'}, '400': {$ref: '#/components/responses/BadRequest'}, '401': {$ref: '#/components/responses/Unauthorized'}, '403': {$ref: '#/components/responses/Forbidden'}, '404': {$ref: '#/components/responses/NotFound'}, '409': {$ref: '#/components/responses/Conflict'}, '422': {$ref: '#/components/responses/ValidationError'}, '500': {$ref: '#/components/responses/InternalServerError'}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    StudentId: {name: studentId, in: path, required: true, schema: {$ref: '#/components/schemas/Id'}}
    ExamId: {name: examId, in: path, required: true, schema: {$ref: '#/components/schemas/Id'}}
...
    Page: {name: page, in: query, schema: {type: integer, minimum: 1, default: 1}}
...
  schemas:
...
    StudentCreate: {type: object, required: [name,currentStageId,currentGradeId], properties: {name: {type: string}, birthday: {type: [string,'null'], format: date}, school: {type: [string,'null']}, currentStageId: {$ref: '#/components/schemas/Id'}, currentGradeId: {$ref: '#/components/schemas/Id'}, remark: {type: [string,'null']}}}
    StudentUpdate: {allOf: [{$ref: '#/components/schemas/StudentCreate'}, {type: object, required: [version], properties: {version: {type: integer}}}]}
...
  responses:
...
    BadRequest: {$ref: '#/components/responses/Error'}
    Unauthorized: {$ref: '#/components/responses/Error'}
...
    NotFound: {$ref: '#/components/responses/Error'}
    Conflict: {$ref: '#/components/responses/Error'}
...
    ValidationError: {$ref: '#/components/responses/Error'}
    InternalServerError: {$ref: '#/components/responses/Error'}
...
    StudentList: {description: 成功, content: {application/json: {schema: {allOf: [{$ref: '#/components/schemas/Envelope'}, {type: object, properties: {data: {type: array, items: {$ref: '#/components/schemas/Student'}}}}]}}}}
    StudentOne: {description: 成功, content: {application/json: {schema: {allOf: [{$ref: '#/components/schemas/Envelope'}, {type: object, properties: {data: {$ref: '#/components/schemas/Student'}}}]}}}}
...